Version 1 Objective System
We have created a primitive version of the coin and flag pick up system along with a coin and flag position randomiser that places the coins and flags in a random position around whatever planet they are on. Also fixed some minor bugs with the player space ship and collision that caused the camera to permanently vibrate after the player collided with a planet. Also fixed a bug that caused Mercury to get shot off into the sun instead of orbiting. Finally we also made it so the player wouldn't be able to redirect the orbit of the planets