Version 5 Map Border and Acceleration Animation
The map border: When travelling away from the sun and is further than a pre-decided distance, the speed of the spaceship exponentially decreases, and ultimately come to a point when the spaceship is almost still. When travelling towards the sun, the speed exponentially increases back to normal.
The acceleration animation is that the flame from the thruster changes colour from blue to red while accelerating.