Version 7 Radar Visual
We decide to use a radar ability as the indicator that assist players to locate the objectives. This version we have added the visual effects for that including "a particle spread" when the player presses R key to trigger radar, and "flash" animation of planets in the area of radar.